requirements.txt
openpyxl==3.1.2
Flask==2.3.3
gunicorn==21.2.0
google-api-python-client==2.108.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
python-dotenv==1.0.0